refactor(cli): split cli/main.py into single-responsibility command modules (#2542, task 3)
Pure move of the 3061-line cli/main.py into ten modules; the typer
command surface is byte-identical (verified: same 37 command names,
same groups, same callback behavior):

- context.py     shared typer app + cfg/state caches, server_url,
                 require_auth/_maybe_none_login, ws_max_size, _err.
                 The single mutable cells (_cfg_cache/_state_cache/
                 _server_override) live here so every command module
                 and the callback write through one module attribute.
- authcmds.py    login/logout/status + account self-service
- workspaces.py  ls/create/dup/rm/members/restart/stop/start/export/
                 import + workspace_status/short_id + the settings
                 editor helpers (_parse_env_list/_build_settings/
                 _prompt/_SENTINEL) shared with edit
- edit.py        the edit command
- shellcmd.py    shell + agent-forwarding + consent-popup helpers
- monitor.py     monitor + _dispatch_monitor_event/monitor_backoff/
                 monitor_connection/monitor_run/refresh_token_threaded
- sandboxcmd.py  sandbox + sandbox_setup/sandbox_setup_only
- terminals.py   terminal ls + window/terminal share/unshare +
                 workspace share/unshare
- execsync.py    exec/sync/images
- admin.py       admin users+invitations, volumes, consent-decide
- main.py        composition + re-export shim (app callback, TUI
                 launcher, main()) so klangk.cli.main.X keeps
                 resolving for callers and patch targets

Command modules call shared context helpers as context.<fn>() (module
attribute) so a single patch of klangk.cli.context.<fn> affects every
caller; the CLI subpackage isolation rule is preserved (no server
imports).

Test adjustments are patch-target path fixes only: monkeypatch/patch
targets moved to the defining module (context for caches/_client/_state/
server_url/_server_override; sandboxcmd for sandbox_setup*/exec_on_ws/
asyncio.run; monitor for monitor_*/refresh_token/local_login/
_server_mode_is_none; shellcmd for ws_shell/popup helpers/asyncio.run;
execsync for ws_exec; authcmds for do_logout/account; admin for
consent-decide Console). No assertions or test logic changed.

4970 passed, coverage 100%.
