+++ src/klangk/klangk/cli/authcmds.py
@@ -1,0 +1,231 @@
+"""Session commands: login, logout, status.
+
+Part of the #2542 split of the former 3061-line
+``klangk/cli/main.py``; commands and helpers live in
+single-responsibility modules, all imported back into
+``klangk.cli.main`` for backwards compatibility.
+"""
+
+from __future__ import annotations
+
+import sys
+from pathlib import Path
+
+import httpx
+import typer
+from rich.console import Console
+from rich.prompt import Confirm, Prompt
+from rich.table import Table
+
+from . import account
+from .auth import login, logout as do_logout
+from .client import decode_token_claims
+from . import context
+
+
+@context.app.command("login")
+def login_cmd(
+    server: str = typer.Argument(..., help="Server alias or URL"),
+    user: str | None = typer.Argument(None, help="User (email or handle)"),
+    password_file: str | None = typer.Option(
+        None,
+        "--password-file",
+        help="Read password from file (use - for stdin)",
+    ),
+) -> None:
+    """Authenticate with a Klangk server."""
+    cfg = context._cfg()
+    resolved_url = cfg.resolve_server(server)
+    # Default user from config if not provided on command line
+    email = user or cfg.get_user(resolved_url)
+    password = None
+    if password_file is not None:
+        if password_file == "-":
+            password = sys.stdin.readline().rstrip("\n")
+        else:
+            password = Path(password_file).read_text().strip()
+    login(resolved_url, email=email, password=password)
+
+
+@context.app.command()
+def logout(
+    server: str | None = typer.Argument(None, help="Server alias or URL"),
+) -> None:
+    """Clear stored credentials."""
+    if server is not None:
+        resolved_url = context._cfg().resolve_server(server)
+    else:
+        active = context._state().active_server
+        if active is None:
+            context._err.print(
+                "[red]No active server[/red] — pass a server argument"
+                " or log in first."
+            )
+            raise typer.Exit(code=1)
+        resolved_url = active
+    do_logout(resolved_url)
+
+
+@context.app.command()
+def status(
+    plain: bool = typer.Option(False, "--plain", help="Plain text output"),
+) -> None:
+    """Show connection info (server, user, admin status)."""
+    # status works even with no active server (unlike other commands).
+    url = context._server_override or context._state().active_server
+    state = context._state()
+    token = state.get_token(url) if url else None
+    email = state.get_email(url) if url else None
+    user_id = decode_token_claims(token).get("sub") if token else None
+    # Admin status comes from /my-permissions (the canonical source the
+    # frontend uses for isAdmin). Best-effort: if the probe fails (offline,
+    # token expired, old server without /admin in the static set) status
+    # still reports everything else rather than erroring out.
+    is_admin: bool | None = None
+    if token:
+        try:
+            client = context._client()
+            resp = client.get("/api/v1/my-permissions")
+            client.check_auth(resp)
+            if resp.status_code == 200:
+                perms = resp.json().get("permissions", {})
+                is_admin = "*" in perms.get("/admin", [])
+        except Exception:
+            is_admin = None
+    if plain:
+        print(f"server={url or '(none)'}")
+        if token:
+            print(f"user={email or 'unknown'}")
+            print(f"user_id={user_id or 'unknown'}")
+            print("status=logged_in")
+            if is_admin is not None:
+                print(f"admin={'yes' if is_admin else 'no'}")
+        else:
+            print("status=not_logged_in")
+        return
+    console = Console()
+    table = Table(show_header=False, box=None, pad_edge=False)
+    table.add_column(style="bold")
+    table.add_column()
+    table.add_row("Server", url or "(none)")
+    if token:
+        table.add_row("User", email or "unknown")
+        table.add_row("User ID", user_id or "unknown")
+        table.add_row("Status", "[green]logged in[/green]")
+        if is_admin:
+            table.add_row("Admin", "[green]yes[/green]")
+        elif is_admin is False:
+            table.add_row("Admin", "no")
+    else:
+        table.add_row("Status", "[yellow]not logged in[/yellow]")
+    console.print(table)
+
+
+# ---------------------------------------------------------------------
+# Account self-service (change password / handle / email)
+# ---------------------------------------------------------------------
+
+account_app = typer.Typer(
+    help="Account self-service: change your password, handle, or email."
+)
+context.app.add_typer(account_app, name="account")
+
+
+@account_app.command("show")
+def account_show() -> None:
+    """Show your current handle and email."""
+    context.require_auth()
+    me = context._client().get_me()
+    handle = me.get("handle") or "(none)"
+    email = me.get("email") or "(unknown)"
+    Console().print(
+        f"Email:  [bold]{email}[/bold]\nHandle: [bold]@{handle}[/bold]"
+    )
+
+
+@account_app.command("passwd")
+def account_passwd() -> None:
+    """Change your password."""
+    context.require_auth()
+    url = context.server_url()
+    client = context._client()
+    current = Prompt.ask("[bold]Current password[/bold]", password=True)
+    new = Prompt.ask("[bold]New password[/bold]", password=True)
+    confirm = Prompt.ask("[bold]Confirm new password[/bold]", password=True)
+    if new != confirm:
+        context._err.print("[red]Passwords do not match[/red]")
+        raise typer.Exit(code=1)
+    min_len = account.password_min_length(url)
+    if len(new) < min_len:
+        context._err.print(
+            f"[red]Password must be at least {min_len} characters[/red]"
+        )
+        raise typer.Exit(code=1)
+    try:
+        client.change_password(current, new)
+    except httpx.HTTPStatusError as exc:
+        # The server surfaces a 401 for a wrong current password and a 400
+        # for policy violations (e.g. too short) — the detail is printed
+        # verbatim. change-password doesn't itself trigger the /auth/login
+        # brute-force lockout; a future global rate limit (429) would show
+        # up here as a raw HTTP error until given dedicated handling.
+        context._err.print(f"[red]{exc}[/red]")
+        raise typer.Exit(code=1) from exc
+    # Success goes to stdout (scripting-friendly: `klangk account passwd &&
+    # ...`); errors above go to stderr via context._err.
+    Console().print("[green]Password updated.[/green]")
+
+
+@account_app.command("handle")
+def account_handle() -> None:
+    """Change your handle (requires password confirmation)."""
+    context.require_auth()
+    client = context._client()
+    current = client.get_me().get("handle") or ""
+    new = Prompt.ask("[bold]New handle[/bold]").strip()
+    err = account.validate_handle(new)
+    if err:
+        context._err.print(f"[red]{err}[/red]")
+        raise typer.Exit(code=1)
+    if not Confirm.ask(
+        f"Change your handle from @{current} to @{new}?", default=False
+    ):
+        context._err.print("[yellow]Cancelled.[/yellow]")
+        raise typer.Exit(code=0)
+    password = Prompt.ask("[bold]Password (to confirm)[/bold]", password=True)
+    try:
+        client.change_handle(new, password)
+    except httpx.HTTPStatusError as exc:
+        context._err.print(f"[red]{exc}[/red]")
+        raise typer.Exit(code=1) from exc
+    Console().print(f"[green]Handle updated to @{new}.[/green]")
+
+
+@account_app.command("email")
+def account_email() -> None:
+    """Change your email (requires password confirmation)."""
+    context.require_auth()
+    url = context.server_url()
+    client = context._client()
+    new = Prompt.ask("[bold]New email[/bold]").strip()
+    err = account.validate_email(new)
+    if err:
+        context._err.print(f"[red]{err}[/red]")
+        raise typer.Exit(code=1)
+    password = Prompt.ask("[bold]Password (to confirm)[/bold]", password=True)
+    try:
+        client.change_email(new, password)
+    except httpx.HTTPStatusError as exc:
+        context._err.print(f"[red]{exc}[/red]")
+        raise typer.Exit(code=1) from exc
+    # The JWT subject is the user id, so the cached token stays valid; only
+    # the key it's filed under changes. Re-key it rather than dropping it.
+    state = context._state()
+    old = state.get_email(url)
+    if old is not None and old != new:
+        state.rename_user(url, old, new)
+        state.save()
+    Console().print(
+        "[green]Email updated.[/green] Check your inbox to verify the new"
+        " address."
+    )
